Platform/RaspberyyPi fix Logo background
The Logo background wasn't actually black. It's some kind
of dark grey. It looks awful on most of my LCDs. Sadly,
that Logo.eps doesn't actually seem to open anywhere
(not on Linux, not on macOS), but fortunately we can
convert to an easy to fix format (XPM) and edit the colors
away.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>